Added recursive metadata extraction and took care of some todos in the APIs
I added getRecursiveMetadata, which recursively collects metadata from a folder downwards and writes it to songs.json. I also took care of some todos in the APIs. Specifically, I mainly added some cross platform handling with writeMetadata so it works on Mac and likely Linux since it uses the same command. In addition to this, I also edited the playlist functions to not require the .json extension(and also create the playlist folder if it doesn't exist in appdata).