Potential fix for a bug where settings.json was appearing in musicplayer/ on startup
On line 94, we can see that fsInit and setPath are called. Theoretically, this should work since setPath is called after fsInit, but since fsInit is asynchronous, it gets called right after all synchronous actions, like setPath, are called. As a result, when setPath calls getSettings, storagePath is still an empty string, and since settings.json doesn't exist in musicplayer/, it gets created. To resolve this, I simply added an await in front of fsInit. Hopefully this doesn't break anything lol