Fixed playSongAPICalls.js
Also, playSong has a new parameter, time. This is for the duration of the song in seconds. The reason why we have this is because we're planning on cutting the input short using -loglevel quiet -stats and front end already called the metadata reading functions, so they can pass it in here